move: main::compile to compile::compilef
