+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod sub_commands;
+pub mod compile;
